refactoring and types update
