+++ app/src/app/(components)/(MainPage)/MainPage.tsx
@@ -16,10 +16,8 @@ export const MainPage = () => {
   return (
     <main>
       <TopToolbar />
-      {/* <CleanButton changeCleanButton={ isPressed => setIsPressed(isPressed)} isPressed={isPressed}/> */}
       <div className="map-container" style={{ position: 'relative' }}>
         <OpenStreetMap zoomLvl={13} />
-        {/* <StartButton /> */}
       </div>
     </main>
   )
