+++ app/src/app/(components)/(OpenStreetMap)/components/SetViewOnClick.tsx
@@ -1,0 +1,18 @@
+import { useEffect } from 'react'
+import { useMap } from 'react-leaflet'
+
+type SetViewOnClickProps = {
+  zoomLvl: number
+}
+
+export const SetViewOnClick = ({ zoomLvl }: SetViewOnClickProps) => {
+  const map = useMap()
+
+  useEffect(() => {
+    map.flyTo(map.getCenter(), zoomLvl, {
+      animate: true,
+      duration: 1, // in seconds
+    })
+  })
+  return null
+}
