+++ app/src/app/(components)/(PolygonEditor)/images/spritesheet-2x.png
