+++ app/src/app/(components)/(PolygonEditor)/images/spritesheet.png
