+++ app/src/app/(components)/(PolygonEditor)/styles/style.css
